requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
websockets==3.4
discord==0.0.2
discord.py==0.16.12
python-dotenv==0.8.2